+++ src/features/view-chapter/components/ReadingContent.tsx
@@ -21,7 +21,7 @@ type Props = {
 };
 
 const ReadingContent: React.FC<Props> = ({ bismiPre, pages }) => {
-  const { id } = useParams<{ id: string }>();
+  const { chapterNo } = useParams<{ chapterNo: string }>();
   const {
     isLoading,
     data: versesUthmaniData,
@@ -30,7 +30,7 @@ const ReadingContent: React.FC<Props> = ({ bismiPre, pages }) => {
     isFetchingNextPage,
     hasNextPage,
   } = useVersesUthmani({
-    chapterId: parseInt(id),
+    chapterId: parseInt(chapterNo),
     pages,
   });